fix permission error bug

+++ src/explorepy/__init__.py
@@ -2,22 +2,27 @@ import os
 import shutil
 from sys import platform as _platform
 
-parent_directory = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir))
+sdk_exists = False
+for file in os.listdir(os.path.dirname(__file__)):
+    if '_exploresdk' in file:
+        sdk_exists = True
 
-files = (file for file in os.listdir(parent_directory)
-         if os.path.isfile(os.path.join(parent_directory, file)))
+if not sdk_exists:
+    parent_directory = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir))
+    files = (file for file in os.listdir(parent_directory)
+             if os.path.isfile(os.path.join(parent_directory, file)))
 
-if _platform == "linux" or _platform == "linux2":
-    for file in files:
-        if '_exploresdk' in file:
-            full_path = os.path.join(parent_directory, file)
-            shutil.copy(full_path, os.path.dirname(__file__))
+    if _platform == "linux" or _platform == "linux2":
+        for file in files:
+            if '_exploresdk' in file:
+                full_path = os.path.join(parent_directory, file)
+                shutil.copy(full_path, os.path.dirname(__file__))
 
-elif _platform == "win32" or _platform == "win64":
-    for file in files:
-        if '_exploresdk' in file:
-            full_path = os.path.join(parent_directory, file)
-            shutil.copy(full_path, os.path.dirname(__file__))
+    elif _platform == "win32" or _platform == "win64":
+        for file in files:
+            if '_exploresdk' in file:
+                full_path = os.path.join(parent_directory, file)
+                shutil.copy(full_path, os.path.dirname(__file__))
 
 
 from .explore import Explore
